fixed the design
