+++ LoC/src/ru/nsu/ccfit/izhitsky/task1/MyComplexFilterFactory.java
@@ -1,7 +1,8 @@
 package ru.nsu.ccfit.izhitsky.task1;
 
-import java.lang.reflect.Method;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class MyComplexFilterFactory
@@ -28,38 +29,60 @@ public class MyComplexFilterFactory
 
 	public MyFilter toCreate(String configLine) //e.g. "&(.java <100)"
 	{
-		char filterType = configLine.charAt(0); //e.g. "&"
+		char filterTypeChar = configLine.charAt(0); //e.g. "&"
 		String filterParameter = configLine.substring(1); //e.g. "(.java <100)"
-		if (!factoryMap.containsKey(filterType))
+
+		if (!factoryMap.containsKey(filterTypeChar))
 		{
 			//TODO: throw something
-			System.err.println("Invalid symbol used in configuration file : " + configLine);
+			System.err.println("ERROR: an invalid symbol was used in the configuration file : " + configLine);
 			System.exit(1);
 		}
+
+		String filterType = factoryMap.get(filterTypeChar);
+		MyFilter[] filters = toParseString(factoryMap.get(filterType)); //e.g. "&" -> "AndFilter"
+
 		try
 		{
-			String selectedFilter = factoryMap.get(filterType); //e.g. "&" -> "AndFilter"
-			switch (selectedFilter)
+			switch (filterType)
 			{
 				case "AndFilter":
-					return new AndFilter(parse(configLine));
+					return new AndFilter(filters);
 					break;
 				case "OrFilter":
-					return new OrFilter(parse(configLine));
+					return new OrFilter(filters);
 					break;
 				case "NotFilter":
-					MyFilter theFilter = null;
+					MyFilter theFilter;
 					if (configLine.charAt(0) == '(')
 					{
-						theFilter = MyComplexFilterFactory.toCreate(configLine.substring(1, configLine.length() - 1));
+						MyComplexFilterFactory tempFactory = new MyComplexFilterFactory();
+						theFilter = tempFactory.toCreate(configLine.substring(1, configLine.length() - 1));
 					}
 					else
 					{
-						theFilter = MyComplexFilterFactory.toCreate(configLine);
+						MyComplexFilterFactory tempFactory = new MyComplexFilterFactory();
+						theFilter = tempFactory.toCreate(configLine);
 					}
 					return new NotFilter(theFilter);
+					break;
+				case "ExtensionFilter":
+					//TODO
+					//return new ExtensionFilter(configLine);
+					break;
+				case "TimeLessFilter":
+					//TODO
+					break;
+				case "TimeGreaterFilter":
+					//TODO
+					break;
+				default:
+					//TODO: throw
+					break;
 			}
 		}
+
+
 		catch (Exception e)
 		{
 			e.printStackTrace();
@@ -67,4 +90,48 @@ public class MyComplexFilterFactory
 
 		return null;
 	}
+
+
+	private static MyFilter[] toParseString(String theString_)
+	{
+		String theString = theString_.substring(1, theString_.length() - 1) + ' ';
+		List<String> filterList = new ArrayList<>();
+		String temp = "";
+		int bracketsCounter = 0;
+
+		for (int i = 0; i < theString.length(); i++)
+		{
+			char c = theString.charAt(i);
+			if (c == '(')
+			{
+				bracketsCounter++;
+			}
+			else if (c == ')')
+			{
+				bracketsCounter--;
+			}
+			else if (c == ' ' && bracketsCounter == 0)
+			{
+				filterList.add(temp);
+				temp = "";
+				continue;
+			}
+			temp = temp + c;
+		}
+
+		if (bracketsCounter != 0)
+		{
+			//TODO: throw something
+		}
+
+		List<MyFilter> filtersList = new ArrayList<>();
+
+		for (String theFilter : filterList)
+		{
+			filtersList.add(MyComplexFilterFactory.toCreate(theFilter));
+		}
+
+		return filtersList.toArray(new MyFilter[filtersList.size()]);
+	}
+
 }
